+++ core/src/main/java/io/grpc/internal/ForwardingClientStreamTracer.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 The gRPC Authors
+ * Copyright 2021 The gRPC Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.grpc.util;
+package io.grpc.internal;
 
 import com.google.common.base.MoreObjects;
 import io.grpc.Attributes;
@@ -23,15 +23,12 @@ import io.grpc.ExperimentalApi;
 import io.grpc.Metadata;
 import io.grpc.Status;
 
-/**
- * Delegation class for ClientStreamTracer.
- *
- * @deprecated Use {@link io.grpc.internal.ForwardingClientStreamTracer} instead.
- */
 @ExperimentalApi("https://github.com/grpc/grpc-java/issues/2861")
-@Deprecated
 public abstract class ForwardingClientStreamTracer extends ClientStreamTracer {
-  /** Returns the underlying {@code ClientStreamTracer}. */
+
+  /**
+   * Returns the underlying {@code ClientStreamTracer}.
+   */
   protected abstract ClientStreamTracer delegate();
 
   @Override
